requirements.txt
fastapi
uvicorn
google_generativeai
requests
langchain
langchain-community
python-dotenv
pydantic